Fetch: Add qualifier checking
Allows us to check whether the qualifiers supplied by the request are
supported by any of the fetchers specified. In order to do this, I've
created a "Fetcher" interface extending the "remoteasset.FetchServer"
interface, which has an additional method to check whether a set of
qualifier names are supported, returning any that are not.

This allows us to discover whether the request is supported by the
server, and propagate the details back to the ValidatingFetcher at the
front to raise an error.

I've also implemented a Set type for qualifier names.