Improve configuration and remove asset blob access
The fetchers and pushers now use blobstore.BlobAccess to interact with
the action cache and content addressable storage.

These are defined in a `blobstore` block in the
`ApplicationConfiguration` rather than both the fetcher and pusher being
able to specify their own gRPC endpoints for the two.

As a result of these changes the asset blob access is no longer required
and has been removed