AC: Use RootDirectoryDigest for Directories
bb-remote-asset does some cleverness to pop `Directory` objects into the
OutputDirectories field of the ActionResult for semantic purposes. This
commit cleans up that cleverness to allow blobs that look a bit like
directories to be uploaded as asset.

Prior to this commit we detected whether an object is a `Directory` by
fetching it from CAS and attempting to unmarshal it into a `Directory`
message. This commit requires all assets uploaded via `PushDirectory`
are `Directories` (and treats assets uploaded via `PushBlob` as blobs).

We still must convert Directories to a Tree to generate the `TreeDigest`
needed by the `ActionResult`, so the action cache asset store will fetch
from the CAS when putting a new directory. (The ActionResult needs a
`TreeDigest` to satisfy bb-storage's completeness checking.)

If the client decides to use `PushDirectory` to push something that
isn't a `Directory` (which isn't explicitly banned by the spec), then we
will
error. The client will need to use `PushBlob` for these assets instead.

Fixes https://github.com/buildbarn/bb-remote-asset/issues/33
Closes https://github.com/buildbarn/bb-remote-asset/pull/34