Tidy up action cache asset store
The Action Cache Asset Store is a bit hard to work with at the moment.  Lots of
logic is included in the `Get` and `Put` methods that can be extracted and
shared, while also clarifying what's really going on in these methods.

I've also modified some of the utility functions to be more generic which has
removed a load of boilerplate.