fetcher: Reject requests without any checksum
No checksum being sent which means that the
request is not guaranteed reproducible. Bazel's downloader can be used
as a general general purpose GET client, and many rulesets abuse this
(rules_oci notably) to fetch things like credentials and things that are
not reproducible and are dangerous to cache.

However, these things generally lack any checksum (as they downloading
dynamic content), so if we require all asset fetch requests to have a
checksum, we can filter out these requests and allow the user to have
the option to fallback to the local downloader for these requests.