Support DigestFunction in messages
The Remote Asset API now supports specifying the Digest Function to use in all
messages.  Let's implement support for that, so that we can avoid breaking users
that want to use a specific digest function.

In doing this, I've made some other modifications, which also improve
correctness.

1. We now pass digest.Functions into the Asset Store, rather than
   digest.InstanceNames.  This lets us use the passed digest function instead of
   guessing/just using SHA256.
2. We use the bb-storage digest/buffer APIs everywhere, instead of using
   REAPI objects and converting
3. Thanks to (2), we now only need a single ProtoSerialise method, that converts
   a proto to a Buffer and Digest.

In addition to being more correct, this has hugely simplified things in the
Remote Execution fetcher and Action Cache asset store.