HTTP Fetcher: Disentangle checksum.sri from Digest
When we implemented HTTP fetching, there was no digest_function field in the
requests, so we used checksum.sri to choose one.  The API has now changed to
explicitly include a digest_function, so using the checksum.sri is actually
incorrect behaviour -- if a client sets digest_function to SHA256, and specifies
a SHA512 checksum, we should validate using the SHA512 checksum, but return a
SHA256 digest.

This change in API may need a wider rework of how checksum validation works.  We
should probably add a decorator fetcher that does this, pulling the object from
CAS based on the inner fetcher's response, and validating that.  This would be
more correct, but at the expense of additional network I/O.