Support DigestFunction in the Remote Asset API (#57)
* Support DigestFunction in messages

The Remote Asset API now supports specifying the Digest Function to use in all
messages.  Let's implement support for that, so that we can avoid breaking users
that want to use a specific digest function.

In doing this, I've made some other modifications, which also improve
correctness.

1. We now pass digest.Functions into the Asset Store, rather than
   digest.InstanceNames.  This lets us use the passed digest function instead of
   guessing/just using SHA256.
2. We use the bb-storage digest/buffer APIs everywhere, instead of using
   REAPI objects and converting
3. Thanks to (2), we now only need a single ProtoSerialise method, that converts
   a proto to a Buffer and Digest.

In addition to being more correct, this has hugely simplified things in the
Remote Execution fetcher and Action Cache asset store.

* HTTP Fetcher: Disentangle checksum.sri from Digest

When we implemented HTTP fetching, there was no digest_function field in the
requests, so we used checksum.sri to choose one.  The API has now changed to
explicitly include a digest_function, so using the checksum.sri is actually
incorrect behaviour -- if a client sets digest_function to SHA256, and specifies
a SHA512 checksum, we should validate using the SHA512 checksum, but return a
SHA256 digest.

This change in API may need a wider rework of how checksum validation works.  We
should probably add a decorator fetcher that does this, pulling the object from
CAS based on the inner fetcher's response, and validating that.  This would be
more correct, but at the expense of additional network I/O.

* HTTP Fetcher: Fix tests for digest function changes

These changes should make the tests less brittle, because it stops us needlessly
mocking the io.ReadCloser we pass around.  Instead we pass a buffer of the
actual test data.