Authorize push server, rather than asset store
Authorizing the asset store means that anything a fetcher wants to push into the
store must pass the push authorizer.  This isn't what we want, as fetchers
should be able to push into the store with impunity.  Instead we move the
authorization up a layer to wrap the push server.  The fetcher is already
wrapped.